v0.4.1.1

+++ lamw_manager/core/installer/installer.sh
@@ -7,7 +7,6 @@
 #Description: "installer.sh" is part of the core of LAMW Manager. Contains routines for installing LAMW development environment
 #-------------------------------------------------------------------------------------------------#
 
-
 #prepare upgrade
 LAMWPackageManager(){
 	if [ $FLAG_FORCE_ANDROID_AARCH64 = 1 ]; then 
@@ -19,38 +18,29 @@ LAMWPackageManager(){
 			rm "$old_lamw_ide_home"  -rf
 		fi
 
-		for((i=0;i<${#LAZARUS_OLD_STABLE[*]};i++))
-		do
-			local old_lazarus_home=$LAMW4LINUX_HOME/${LAZARUS_OLD_STABLE[i]}
-			if [ -e "$old_lazarus_home" ]; then
-				rm "$old_lazarus_home" -rf
-			fi
+		for((i=0;i<${#OLD_LAZARUS_STABLE_VERSION[*]};i++)); do
+			local old_lazarus_release=lazarus_${OLD_LAZARUS_STABLE_VERSION[i]//\./_}
+			local old_lazarus_home=$LAMW4LINUX_HOME/${old_lazarus_release}
+			[ -e "$old_lazarus_home" ] && rm "$old_lazarus_home" -rf
 		done
 		
-		if [ -e "$OLD_FPC_CFG_PATH" ]; then
-			rm "$OLD_FPC_CFG_PATH"
-		fi
+		[ -e "$OLD_FPC_CFG_PATH" ] && grep  "$ROOT_LAMW" "$OLD_FPC_CFG_PATH" && rm "$OLD_FPC_CFG_PATH"
 	
 		#fixs 0.3.1 to 0.3.2
 
 		for i  in ${!OLD_FPC_SOURCES[*]}; do
-			if [ -e ${OLD_FPC_SOURCES[i]} ]; then
-				rm  -rf ${OLD_FPC_SOURCES[i]}
-			fi
+			[ -e ${OLD_FPC_SOURCES[i]} ] && rm  -rf ${OLD_FPC_SOURCES[i]}
 		done
 
 
-
-		for gradle in ${OLD_GRADLE[*]}
-		do
+		for gradle in ${OLD_GRADLE[*]}; do
 			if [ -e "$gradle" ]; then
-				./$gradle/bin/gradle --stop
+				stopGradleDaemon "$gradle"
 				rm -rf $gradle 
 			fi
 		done
 
-		for ((i=0;i<${#OLD_ANT[*]};i++))
-		do
+		for ((i=0;i<${#OLD_ANT[*]};i++)); do
 			[ -e ${OLD_ANT[i]} ] && rm -rf ${OLD_ANT[i]}
 		done
 
@@ -59,16 +49,12 @@ LAMWPackageManager(){
 		done
 	fi
 
-
-
 }
 getStatusInstalation(){
 	if [  -e $LAMW4LINUX_HOME/lamw-install.log ]; then
 		export LAMW_INSTALL_STATUS=1
 		return 1
 	else 
-		setOldAndroidSDKStatus
-		export NO_GUI_OLD_SDK=1
 		return 0;
 	fi
 }
@@ -82,16 +68,34 @@ installDependences(){
 }
 
 
+
+
+getCompressFile(){
+	local compress_url="$1"
+	local compress_file="$2"
+	local uncompress_command="$3"
+	local before_uncompress="$4"
+	local error_uncompress_msg="${VERMELHO}Error:${NORMAL} corrupt/unsupported file"
+	local initial_msg="Please wait, extracting ${NEGRITO}$compress_file${NORMAL} ..." 
+	Wget $compress_url
+	if [ -e $compress_file ]; then
+		printf "%s" "$initial_msg"	
+		[ "$before_uncompress" != "" ] &&  eval "$before_uncompress"
+		$uncompress_command
+		check_error_and_exit "$error_uncompress_msg"
+		printf  "%s\n" "${FILLER:${#compress_file}}${VERDE} [OK]${NORMAL}"
+		rm $compress_file
+	fi
+}
+
 getJDK(){
 	checkJDKVersionStatus
 	if [ $JDK_STATUS = 1 ]; then 
 		[ -e "$OLD_JAVA_HOME" ] && rm -rf "$OLD_JAVA_HOME"
 		changeDirectory "$ROOT_LAMW/jdk"
 		[ -e "$JAVA_HOME" ] && rm -r "$JAVA_HOME"
-		Wget "$JDK_URL"
-		tar -zxvf "$JDK_TAR"
-		mv "$JDK_FILE" "${JDK_VERSION_FOLDER}"
-		[ -e "$JDK_TAR" ] && rm "$JDK_TAR"
+		getCompressFile "$JDK_URL" "$JDK_TAR" "tar -zxf $JDK_TAR"
+		mv "$JDK_FILE" "${JDK_VERSION_DIR}"
 	fi
 }
 
@@ -118,11 +122,19 @@ getFromGit(){
 			local git_param=(pull origin $git_branch )
 		fi
 		changeDirectory "$git_src_dir"
+		git config pull.ff only
+
 		git ${git_param[@]}
 		if [ $? != 0 ]; then
-			git_param=(reset --hard)
+		local git_reset_param=(reset --hard)
+			git ${git_reset_param[@]}
 			git ${git_param[@]}
-			check_error_and_exit "possible network instability!! Try later!"
+			if [ $? != 0 ]; then
+				changeDirectory .. 
+				rm -rf $git_src_dir
+				echo "possible network instability!! Try later!"
+				exit 1
+			fi
 		fi
 	fi
 }
@@ -151,24 +163,13 @@ getFPCStable(){
 
 	cd "$LAMW4LINUX_HOME/usr"
 	if  [ ! -e "$FPC_LIB_PATH" ]; then
-		echo "doesn't exist $FPC_LIB_PATH"
-		Wget $FPC_DEB_LINK
-		if [ -e "$FPC_DEB" ]; then
-			local tmp_files=(
-				data.tar.xz
-				control.tar.gz
-				debian-binary
-				"$FPC_DEB"
-			)
-			ar x "$FPC_DEB"
-			if [ -e data.tar.xz ]; then 
-				tar -xvf data.tar.xz 
-				rm -rf $LAMW4LINUX_HOME/usr/local
-				[ -e $LAMW4LINUX_HOME/usr/usr ] && mv $LAMW4LINUX_HOME/usr/usr/ $LAMW4LINUX_HOME/usr/local
-			fi
-			for i in ${!tmp_files[@]}; do  
-				if [ -e ${tmp_files[i]} ]; then rm ${tmp_files[i]} ; fi
-			done	
+
+		getCompressFile "$FPC_DEB_LINK" "$FPC_DEB" "ar x $FPC_DEB data.tar.xz"
+		if [ -e data.tar.xz ]; then 
+			tar -xf data.tar.xz 
+			rm -rf $LAMW4LINUX_HOME/usr/local
+			[ -e $LAMW4LINUX_HOME/usr/usr ] && mv $LAMW4LINUX_HOME/usr/usr/ $LAMW4LINUX_HOME/usr/local
+			rm data.tar.xz
 		fi
 		export PPC_CONFIG_PATH=$FPC_LIB_PATH
 
@@ -179,14 +180,22 @@ getFPCStable(){
 getFPCSourcesTrunk(){
 	mkdir -p $FPC_TRUNK_SOURCE_PATH
 	changeDirectory $FPC_TRUNK_SOURCE_PATH
-	getFromSVN "$FPC_TRUNK_URL" "$FPC_TRUNK_SVNTAG"
+	if [ ! -e $FPC_TRUNK_SVNTAG ]; then
+		local url_fpc_src="https://sourceforge.net/projects/freepascal/files/Source/${_FPC_TRUNK_VERSION}/fpc-${_FPC_TRUNK_VERSION}.source.tar.gz"
+		local tar_fpc_src="fpc-${_FPC_TRUNK_VERSION}.source.tar.gz"
+		local untar_fpc_src="tar -zxf $tar_fpc_src"
+		local fpc_src_file="fpc-${_FPC_TRUNK_VERSION}"
+		getCompressFile "$url_fpc_src" "$tar_fpc_src" "$untar_fpc_src"
+		mv $fpc_src_file $FPC_TRUNK_SVNTAG
+	fi
+	#getFromSVN "$FPC_TRUNK_URL" "$FPC_TRUNK_SVNTAG"
 	parseFPCTrunk
 }
 
 #get Lazarus Sources
 getLazarusSources(){
 	changeDirectory $LAMW4LINUX_HOME
-	getFromSVN "$LAZARUS_STABLE_SRC_LNK" "$LAZARUS_STABLE"
+	getFromGit "$LAZARUS_STABLE_SRC_LNK" "$LAZARUS_STABLE" "$LAZARUS_STABLE"
 }
 
 #GET LAMW FrameWork
@@ -200,8 +209,8 @@ getLAMWFramework(){
 
 
 AntTrigger(){
-	if [ $OLD_ANDROID_SDK = 0 ] && [ -e "$ANDROID_SDK/tools/ant" ]; then 
-		mv  "$ANDROID_SDK/tools/ant" "$ANDROID_SDK/tools/.ant"
+	if [ $OLD_ANDROID_SDK = 0 ] && [ -e "$ANDROID_SDK_ROOT/tools/ant" ]; then 
+		mv  "$ANDROID_SDK_ROOT/tools/ant" "$ANDROID_SDK_ROOT/tools/.ant"
 	fi
 }
 
@@ -213,12 +222,10 @@ getAnt(){
 	if [ ! -e "$ANT_HOME" ]; then
 		MAGIC_TRAP_INDEX=0 # preperando o indice do arquivo/diretório a ser removido
 		trap TrapControlC  2
-		Wget $ANT_TAR_URL
 		MAGIC_TRAP_INDEX=1
-		tar -xvf "$ANT_TAR_FILE"
+		getCompressFile "$ANT_TAR_URL" "$ANT_TAR_FILE" "tar -xf $ANT_TAR_FILE"
 	fi
 
-	[ -e  $ANT_TAR_FILE ] && rm $ANT_TAR_FILE
 }
 
 getGradle(){
@@ -226,68 +233,49 @@ getGradle(){
 	if [ ! -e "$GRADLE_HOME" ]; then
 		MAGIC_TRAP_INDEX=2 #Set arquivo a ser removido
 		trap TrapControlC  2 # set armadilha para o signal2 (siginterrupt)
-		Wget $GRADLE_ZIP_LNK
-		MAGIC_TRAP_INDEX=3
-		unzip -o  $GRADLE_ZIP_FILE
-	fi
-
-	if [ -e  $GRADLE_ZIP_FILE ]; then
-		rm $GRADLE_ZIP_FILE
+		getCompressFile "$GRADLE_ZIP_LNK" "$GRADLE_ZIP_FILE" "unzip -o -q $GRADLE_ZIP_FILE" "MAGIC_TRAP_INDEX=3"
 	fi
 }
 
-getNDK(){
-	[ $OLD_ANDROID_SDK = 0 ] && return 
-
-	changeDirectory "$ANDROID_SDK"
-	if [ -e  "$ANDROID_SDK/ndk-bundle" ]; then 
-		for i in ${!OLD_NDK_VERSION_STR[*]}; do
-			grep ${OLD_NDK_VERSION_STR[i]} "$ANDROID_SDK/ndk-bundle/source.properties" > /dev/null
-			[ $? = 0 ] && rm -rf $ANDROID_SDK/ndk-bundle && break
-		done
-	fi
-
-	if [ ! -e ndk-bundle ] ; then
-		MAGIC_TRAP_INDEX=6
-		trap TrapControlC 2 
-		Wget $NDK_URL	
-		MAGIC_TRAP_INDEX=7
-		unzip -o  $NDK_ZIP
-		MAGIC_TRAP_INDEX=-1
-		mv $NDK_DIR_UNZIP ndk-bundle
-		[ -e $NDK_ZIP ] && rm $NDK_ZIP
-
-	fi
-}
 #Get Gradle and SDK Tools 
 getAndroidSDKTools(){
 	initROOT_LAMW
 	changeDirectory $ROOT_LAMW
-
-	if [ $OLD_ANDROID_SDK = 1 ]; then #mode OLD SDK (2-4 with ant support )
-		SDK_TOOLS_VERSION="r25.2.5"
-		SDK_TOOLS_URL="https://dl.google.com/android/repository/tools_r25.2.5-linux.zip"
-		SDK_TOOLS_ZIP="tools_r25.2.5-linux.zip"
-		SDK_TOOLS_DIR="$ANDROID_SDK/tools"
-	fi
-
-	changeDirectory $ANDROID_SDK
-	if [ ! -e "$SDK_TOOLS_DIR" ];then
-		[ $OLD_ANDROID_SDK = 0 ]  && mkdir -p "$SDK_TOOLS_DIR" && changeDirectory "$SDK_TOOLS_DIR"
+	changeDirectory $ANDROID_SDK_ROOT
+	
+	if [ ! -e "$CMD_SDK_TOOLS_DIR" ];then
+		mkdir -p "$CMD_SDK_TOOLS_DIR"
+		changeDirectory "$CMD_SDK_TOOLS_DIR"
 		trap TrapControlC  2
 		MAGIC_TRAP_INDEX=4
-		Wget $SDK_TOOLS_URL
-		MAGIC_TRAP_INDEX=5
-		unzip -o  $SDK_TOOLS_ZIP
-		[ $OLD_ANDROID_SDK = 0 ] && mv cmdline-tools latest
-		rm $SDK_TOOLS_ZIP
+		getCompressFile "$CMD_SDK_TOOLS_URL" "$CMD_SDK_TOOLS_ZIP" "unzip -o -q  $CMD_SDK_TOOLS_ZIP" "MAGIC_TRAP_INDEX=5"
+		mv cmdline-tools latest
 	fi
+}
+
+getSDKAntSupportedTools(){
+	initROOT_LAMW
+	changeDirectory $ANDROID_SDK_ROOT
+	SDK_TOOLS_VERSION="r25.2.5"
+	SDK_TOOLS_URL="https://dl.google.com/android/repository/tools_r25.2.5-linux.zip"
+	SDK_TOOLS_ZIP="tools_r25.2.5-linux.zip"
+	SDK_TOOLS_DIR="$ANDROID_SDK_ROOT/tools"
+	if [ ! -e "$SDK_TOOLS_DIR" ];then
+		trap TrapControlC  2
+		MAGIC_TRAP_INDEX=4
+		getCompressFile "$SDK_TOOLS_URL" "$SDK_TOOLS_ZIP" "unzip -o -q $SDK_TOOLS_ZIP"  "MAGIC_TRAP_INDEX=5"
+	fi
+}
+
+getAndroidCmdLineTools(){
+	getSDKAntSupportedTools
+	getAndroidSDKTools
 	AntTrigger
 }
 
 
 runSDKManagerLicenses(){
-	local sdk_manager_cmd="$SDK_TOOLS_DIR/latest/bin/sdkmanager"
+	local sdk_manager_cmd="$CMD_SDK_TOOLS_DIR/latest/bin/sdkmanager"
 	yes | $sdk_manager_cmd ${SDK_LICENSES_PARAMETERS[*]} 
 	if [ $? != 0 ]; then 
 		yes | $sdk_manager_cmd ${SDK_LICENSES_PARAMETERS[*]} 
@@ -295,8 +283,8 @@ runSDKManagerLicenses(){
 	fi
 }
 
-  runSDKManager(){
-	local sdk_manager_cmd="$SDK_TOOLS_DIR/latest/bin/sdkmanager"
+runSDKManager(){
+	local sdk_manager_cmd="$CMD_SDK_TOOLS_DIR/latest/bin/sdkmanager"
 	if [ $FORCE_YES = 1 ]; then 
 		yes | $sdk_manager_cmd $*
 
@@ -314,22 +302,8 @@ runSDKManagerLicenses(){
 	fi
 }
 
-runOldSDKManager(){
-	local sdk_pack="$1"
-	local sdk_pack_path="$2"
-	local sdk_manager_cmd="$ANDROID_SDK/tools/android"
-	local sdk_cmd_extra_params=(update sdk --all --no-ui --filter $sdk_pack  ${SDK_MANAGER_CMD_PARAMETERS2_PROXY[*]})
-	if [ ! -e $sdk_pack_path ]; then 
-		echo "y" |  $sdk_manager_cmd ${sdk_cmd_extra_params[@]}
 
-		if [ ! -e $sdk_pack_path ]; then 
-			echo "y" |  $sdk_manager_cmd ${sdk_cmd_extra_params[@]}
-			[ ! -e "$sdk_pack_path" ] && echo "possible network instability! Try later!" && exit 1
-		fi
-	fi
-}
-
-getSDKAndroid(){
+getAndroidAPIS(){
 	
 	FORCE_YES=1
 	changeDirectory $ANDROID_HOME
@@ -338,7 +312,7 @@ getSDKAndroid(){
 	if [ $#  = 0 ]; then 
 
 		for((i=0;i<${#SDK_MANAGER_CMD_PARAMETERS[*]};i++));do
-			echo "Please wait, downloading ${NEGRITO}${SDK_MANAGER_CMD_PARAMETERS[i]}${NORMAL}\"..."
+			echo "Please wait, downloading ${NEGRITO}${SDK_MANAGER_CMD_PARAMETERS[i]}${NORMAL}..."
 			
 			if [ $i = 0 ]; then 
 				runSDKManager ${SDK_MANAGER_CMD_PARAMETERS[i]} # instala sdk sem intervenção humana 
@@ -354,32 +328,8 @@ getSDKAndroid(){
 	unset FORCE_YES
 }
 
-getOldAndroidSDK(){
-	local sdk_manager_sdk_paths=(
-		"$ANDROID_SDK/platforms/android-$ANDROID_SDK_TARGET"
-		"$ANDROID_SDK/platform-tools"
-		"$ANDROID_SDK/build-tools/$ANDROID_BUILD_TOOLS_TARGET"
-		"$ANDROID_SDK/extras/google/google_play_services"
-		$ANDROID_SDK/extras/{android,google}/m2repository
-		$ANDROID_SDK/extras/google/market_{licensing,apk_expansion}
-	#	"$ANDROID_SDK/build-tools/$GRADLE_MIN_BUILD_TOOLS"
-	)
 
-	if [ -e $ANDROID_SDK/tools/android  ]; then 
-		changeDirectory $ANDROID_HOME
-		if [ $NO_GUI_OLD_SDK = 0 ]; then
-			echo "Please wait..."
-			$ANDROID_SDK/tools/android  update sdk
-		else 
-			for((i=0;i<${#SDK_MANAGER_CMD_PARAMETERS2[*]};i++));do
-				echo "Getting ${NEGRITO}${SDK_MANAGER_CMD_PARAMETERS2[i]}${NORMAL} ..."
-				runOldSDKManager "${SDK_MANAGER_CMD_PARAMETERS2[i]}" "${sdk_manager_sdk_paths[i]}"
-			done 
-		fi
-	fi
-}
-
-RepairOldSDKAndroid(){
+resetAndroidAPIS(){
 	local sdk_manager_fails=(
 		"platforms"
 		"platform-tools"
@@ -388,9 +338,8 @@ RepairOldSDKAndroid(){
 	)
 	echo "${VERMELHO}Warning:${NORMAL} All Android API'S will  be unistalled!"
 	echo "Only the default APIs will be reinstalled!"
-	for ((i=0;i<${#sdk_manager_fails[*]};i++))
-	do
-		local current_sdk_path="${ANDROID_SDK}/${sdk_manager_fails[i]}"
+	for ((i=0;i<${#sdk_manager_fails[*]};i++)); do
+		local current_sdk_path="${ANDROID_SDK_ROOT}/${sdk_manager_fails[i]}"
 		if [ -e $current_sdk_path ]; then
 			rm -rf $current_sdk_path
 		fi
@@ -398,9 +347,8 @@ RepairOldSDKAndroid(){
 
 	setLAMWDeps
 	getAndroidAPIS
-	for ((i=0;i<${#sdk_manager_fails[*]};i++))
-	do
-		local current_sdk_path="${ANDROID_SDK}/${sdk_manager_fails[i]}"
+	for ((i=0;i<${#sdk_manager_fails[*]};i++)); do
+		local current_sdk_path="${ANDROID_SDK_ROOT}/${sdk_manager_fails[i]}"
 		if [ -e $current_sdk_path ]; then
 			chown $LAMW_USER:$LAMW_USER -R $current_sdk_path
 		fi
@@ -409,18 +357,6 @@ RepairOldSDKAndroid(){
 
 }
 
-
-
-#Addd sdk to .bashrc and .profile
-
-
-getAndroidAPIS(){
-	if [ $OLD_ANDROID_SDK = 0 ]; then
-		getSDKAndroid $*
-	else
-		getOldAndroidSDK
-	fi
-}
 Repair(){
 	local flag_need_repair=0 # flag de reparo 
 	local flag_upgrade_lazarus=0
@@ -456,14 +392,6 @@ Repair(){
 	fi
 }
 
-setOldAndroidSDKStatus(){
-	local lamw_install_log_path="$LAMW4LINUX_HOME/lamw-install.log"
-	if [ -e $lamw_install_log_path ] ; then 
-		#grep "OLD_ANDROID_SDK=0" "$lamw_install_log_path"  > /dev/null 
-	 	OLD_ANDROID_SDK=0
-	fi
-}
-
 checkLAMWManagerVersion(){
 	local ret=0
 	[ -e "$LAMW4LINUX_HOME/lamw-install.log" ] && for i  in ${!OLD_LAMW_INSTALL_VERSION[*]};do
@@ -487,8 +415,6 @@ setOldLAMW4LinuxActions(){
 	if [ $1 = 1 ]; then 
 		echo "You need upgrade your LAMW4Linux!"
 		export LAMW_IMPLICIT_ACTION_MODE=0
-		export NO_GUI_OLD_SDK=1
-		LAMWPackageManager
 	else
 		echo "${VERMELHO}Your LAMW development environment was generated by a newer version of LAMW Manager!${NORMAL}"
 		exit 1
@@ -511,10 +437,8 @@ getImplicitInstall(){
 	local lamw_install_log_path="$LAMW4LINUX_HOME/lamw-install.log"
 
 	if [ ! -e "$lamw_install_log_path" ]; then
-		export NO_GUI_OLD_SDK=1
+		return 
 	else
-		setOldAndroidSDKStatus $?
-		
 		grep "Generate LAMW_INSTALL_VERSION=$LAMW_INSTALL_VERSION" "$lamw_install_log_path" 	> /dev/null
 		
 		if [ $? = 0 ]; then
@@ -527,47 +451,52 @@ getImplicitInstall(){
 	fi
 }
 
-#Build lazarus ide
-BuildLazarusIDE(){
-	
-	wrapperParseFPC
-	
-	export PATH=$FPC_TRUNK_LIB_PATH:$FPC_TRUNK_EXEC_PATH:$PATH
-	local make_opts=(
-		"PP=${FPC_TRUNK_LIB_PATH}/ppcx64"
-		"FPC_VERSION=$_FPC_TRUNK_VERSION"
-	)
-
+installLAMWPackages(){
 	local ide_make_cfg_path="$LAMW4_LINUX_PATH_CFG/idemake.cfg"
-
-	[ ! -e "$LAMW_IDE_HOME" ] && ln -sf $LAMW4LINUX_HOME/$LAZARUS_STABLE $LAMW_IDE_HOME # link to lamw4_home directory  
-	[ ! -e "$LAMW4LINUX_EXE_PATH" ] && ln -sf $LAMW_IDE_HOME/lazarus $LAMW4LINUX_EXE_PATH  #link  to lazarus executable
-
-	changeDirectory $LAMW_IDE_HOME
+	local error_lazbuild_msg="${VERMELHO}Error${NORMAL}: Fails on build ${NEGRITO}${LAMW_PACKAGES[i]}${NORMAL} package"
 	
-	[ $# = 0 ] && make clean all  ${make_opts[*]} #build all IDE
-	
-	initLAMw4LinuxConfig
-
-	if [ -e   "$ide_make_cfg_path" ]; then 
+	if [ -e  "$ide_make_cfg_path" ]; then 
 		local current_widget_set="$(grep '\-dLCL.' "$ide_make_cfg_path" | sed 's/-dLCL//g')"
-		if [ "$current_widget_set" != "" ] && [ "$current_widget_set" != "gtk2" ]; then 
-			current_widget_set="--ws=$current_widget_set"
-		else 
-			current_widget_set=""
-		fi
+		[ "$current_widget_set" != "" ] && current_widget_set="--ws=$current_widget_set"
 	fi
-		#build ide  with lamw framework 
-	for((i=0;i< ${#LAMW_PACKAGES[@]};i++))
-	do
-		local lamw_build_opts=(--build-ide= --add-package ${LAMW_PACKAGES[i]} --pcp=$LAMW4_LINUX_PATH_CFG  --lazarusdir=$LAMW_IDE_HOME $current_widget_set)
+
+	#build ide  with lamw framework 
+	for((i=0;i< ${#LAMW_PACKAGES[@]};i++)); do
+		
+		local lamw_build_opts=(
+			--build-ide= --add-package ${LAMW_PACKAGES[i]} 
+			--pcp=$LAMW4_LINUX_PATH_CFG  
+			--lazarusdir=$LAMW_IDE_HOME 
+			$current_widget_set
+		)
 		./lazbuild  ${lamw_build_opts[*]}
-		if [ $? != 0 ]; then
+		
+		if [ $? != 0 ]; then 
 			./lazbuild ${lamw_build_opts[*]}
-			[ $? != 0 ] && { echo "${VERMELHO}Error${NORMAL}: Fails on build ${NEGRITO}${LAMW_PACKAGES[i]}${NORMAL} package" && return ; }
+			[ $? != 0 ] && { echo "$error_lazbuild_msg" && return ; }
 		fi
 	done
 
+}
+#Build lazarus ide
+BuildLazarusIDE(){
+	[ ! -e "$LAMW_IDE_HOME" ] && ln -sf $LAMW4LINUX_HOME/$LAZARUS_STABLE $LAMW_IDE_HOME # link to lamw4_home directory  
+	[ ! -e "$LAMW4LINUX_EXE_PATH" ] && ln -sf $LAMW_IDE_HOME/lazarus $LAMW4LINUX_EXE_PATH  #link  to lazarus executable
+	
+	wrapperParseFPC
+	export PATH=$FPC_TRUNK_LIB_PATH:$FPC_TRUNK_EXEC_PATH:$PATH
+	local error_build_lazarus_msg="${VERMELHO}Fatal error:${NORMAL}Fails in build Lazarus!!"
+	local make_opts=( "clean all" "PP=${FPC_TRUNK_LIB_PATH}/ppcx64" "FPC_VERSION=$_FPC_TRUNK_VERSION" )
+	
+	changeDirectory $LAMW_IDE_HOME
+
+	if [ $# = 0 ]; then 
+	 	make ${make_opts[*]}
+	 	check_error_and_exit "$error_build_lazarus_msg" #build all IDE
+	fi
+	
+	initLAMw4LinuxConfig
+	installLAMWPackages
 	strip lazarus
 	strip lazbuild
 	strip startlazarus
@@ -582,20 +511,19 @@ checkProxyStatus(){
 	fi
 }
 
-
-
 mainInstall(){
+	getFiller
 	checkLAMWManagerVersion > /dev/null
 	initROOT_LAMW
 	installDependences
 	setLAMWDeps
+	LAMWPackageManager
 	checkProxyStatus
 	wrapperParseFPC
 	getJDK
 	getAnt
 	getGradle
-	getAndroidSDKTools
-	getNDK
+	getAndroidCmdLineTools
 	disableTrapActions
 	getAndroidAPIS 
 	getFPCStable
